[Rename] study 모듈 internal 정리
