+++ src/main/kotlin/com/whatever/caro/study/internal/cardlearningstate/CardLearningStateRepository.kt
@@ -1,4 +1,4 @@
-package com.whatever.caro.study.internal
+package com.whatever.caro.study.internal.cardlearningstate
 
 import org.springframework.data.jpa.repository.JpaRepository
