+++ src/main/kotlin/com/whatever/caro/study/internal/studysession/StudySessionRepository.kt
@@ -1,4 +1,4 @@
-package com.whatever.caro.study.internal
+package com.whatever.caro.study.internal.studysession
 
 import org.springframework.data.jpa.repository.JpaRepository
